fix: windows different partition path relative
description: In windows, if localNodeModules and filename is in different partition, path.relative will return the value of filename. then filenameInLocalNodeModules will be true. It is fault.
solved: In this case, the first char of localNodeModules and filename will be different. adding a condition make filenameInLocalNodeModules
In unix, the new condition will always be true, don't break the logic before.